refactor(organisms): merge SfGroupedProduct
