--- packages/vue/src/components/organisms/SfGroupedProduct/SfGroupedProduct.html
@@ -1,8 +1,0 @@
-<div class="glide sf-grouped-product" ref="glide">
-  <div class="glide__track" data-glide-el="track">
-    <ul class="glide__slides">
-      <!-- @slot Slot for Grouped Product Items -->
-      <slot />
-    </ul>
-  </div>
-</div>
